Use publication_date from utag as alternate in scrape and extract; move publication_date_from_utag to utils; add condition if no existing doc is found in scraping
